Add files via upload
This is a simple script written in Python. It performs basic API testing for an endpoint exposed via AWS API Gateway. The test checks the functionality of an API connected to an AWS Lambda function that interacts with DynamoDB to manage a visitor count. The script makes a GET request to the API endpoint to retrieve the current visitor count stored in DynamoDB. After retrieving the visitor count, the script sends a POST request back to the API to update the visitor count.